Rename `HttpServer` to `Http`
Removed -Server from `HttpServer` package and class since
it is not being used as a server.
