+++ src/test/java/http/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package httpserver;
+package http;
 
 import org.junit.Assert;
 import org.junit.Test;
